requirements.txt
fastmcp>=0.2.0
requests>=2.31.0
starlette>=0.31.1
geopy>=2.4.0
python-dotenv>=1.0.0
uvicorn>=0.23.2
